ENHANCEMENTS: - autorole added to give people the DCS role on join or after linkme CHANGES: - Extensions: some changes in startup / process detection of Lardoon, Sneaker, Olympus, SRS BUGFIX: - /unlink did not work

+++ core/data/impl/serverimpl.py
@@ -470,7 +470,9 @@ class ServerImpl(Server):
 
     def _terminate(self) -> None:
         if self.process and self.process.is_running():
-            self.process.kill()
+            self.process.terminate()
+            if self.process.is_running():
+                self.process.kill()
         self.process = None
 
     async def apply_mission_changes(self, filename: Optional[str] = None) -> str:
